refactor(enrollment): update data storage
